requirements.txt
Django==4.2.0
Pillow==9.5.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-ckeditor==6.5.1
whitenoise==6.5.0
